+++ src/collective/liveblog/browser/recent_updates.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
+from App.Common import rfc1123_date
 from collective.liveblog.interfaces import IBrowserLayer
 from collective.liveblog.interfaces import ILiveblog
+from datetime import datetime
 from five import grok
 from plone.memoize import ram
 
@@ -21,6 +23,37 @@ class RecentUpdates(grok.View):
     grok.name('recent-updates')
     grok.require('zope2.View')
     grok.template('recent_updates')
+
+    def _if_modified_since_request_handler(self):
+        """Return a status code of 304 (not modified) if the requested
+        variant has not been modified since the time specified.
+        """
+        header = self.request.get_header('If-Modified-Since', None)
+        if header is not None:
+            # do what RFC 2616 tells to do in case of invalid date
+            # http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html
+            header = header.split(';')[0]
+            try:
+                # parse RFC 1123 format and normalize for comparison
+                mod_since = datetime.strptime(header, '%a, %d %b %Y %H:%M:%S %Z')
+                mod_since = mod_since.strftime('%Y-%m-%d %H:%M:%S')
+            except (TypeError, ValueError):
+                mod_since = None
+            if mod_since is not None:
+                # convert to UTC and normalize for comparison
+                modified = self.context.modified().utcdatetime()
+                modified = modified.strftime('%Y-%m-%d %H:%M:%S')
+                if modified <= mod_since:
+                    self.request.RESPONSE.setStatus(304)  # not modified
+                    return True
+
+    def update(self):
+        last_modified = rfc1123_date(self.context.modified())
+        self.request.RESPONSE.setHeader('Cache-Control', 'public')
+        self.request.RESPONSE.setHeader('Last-Modified', last_modified)
+
+        if self._if_modified_since_request_handler():
+            return ''
 
     def _updates_since_timestamp(self, timestamp):
         """Return the list of micro-updates since the specified timestamp."""