+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 1.0a2 (unreleased)
 ------------------
 
+- Implement handling of ``If-Modified-Since`` request header on ``recent-updates`` view.
+  [hvelarde]
+
 - Markup of time tag used on automatic updates was fixed.
   [hvelarde]